pyproject.toml
[project]
name = "bible_app"
version = "1.0.0"
description = "The basic blog app built in the Flask tutorial."
dependencies = [
    "flask",
    "pyyaml",
    "torch",
    "waitress",
    "numpy",
    "black",
    "pytest"
]

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"